+++ docs/changelog.rst
@@ -15,6 +15,8 @@ Added
 - config component ``WflowConfigComponent``: represents the Wflow configuration TOML file.
 - staticmaps component ``WflowStaticMapsComponent``: represents Wflow static and cyclic data (used to be grid).
 - states component ``WflowStatesComponent``: represents Wflow input states data.
+- geoms component ``WflowGeomsComponent``: represents Wflow staticgeoms data.
+- **write_geoms**: added function arguments ``to_wgs84``  to convert the geometry to WGS84 before writing it to file. PR #432
 
 Changed
 -------